requirements.txt
python-telegram-bot[webhooks]==20.7
psycopg2-binary==2.9.9
Pillow==10.0.1
